+++ src/Symfony/Bridge/Twig/Extension/TranslationExtension.php
@@ -16,7 +16,6 @@ use Symfony\Bridge\Twig\NodeVisitor\TranslationNodeVisitor;
 use Symfony\Bridge\Twig\TokenParser\TransChoiceTokenParser;
 use Symfony\Bridge\Twig\TokenParser\TransDefaultDomainTokenParser;
 use Symfony\Bridge\Twig\TokenParser\TransTokenParser;
-use Symfony\Component\Translation\LegacyTranslatorTrait;
 use Symfony\Component\Translation\TranslatorInterface as LegacyTranslatorInterface;
 use Symfony\Contracts\Translation\TranslatorInterface;
 use Symfony\Contracts\Translation\TranslatorTrait;
@@ -29,6 +28,8 @@ use Twig\TwigFilter;
  * Provides integration of the Translation component with Twig.
  *
  * @author Fabien Potencier <fabien@symfony.com>
+ *
+ * @final since Symfony 4.2
  */
 class TranslationExtension extends AbstractExtension
 {
@@ -38,21 +39,28 @@ class TranslationExtension extends AbstractExtension
         trans as private doTrans;
     }
 
-    use LegacyTranslatorTrait {
-        transChoice as private doTransChoice;
-    }
-
     private $translator;
     private $translationNodeVisitor;
 
-    public function __construct(TranslatorInterface $translator = null, NodeVisitorInterface $translationNodeVisitor = null)
+    /**
+     * @param TranslatorInterface|null $translator
+     */
+    public function __construct($translator = null, NodeVisitorInterface $translationNodeVisitor = null)
     {
+        if (null !== $translator && !$translator instanceof LegacyTranslatorInterface && !$translator instanceof TranslatorInterface) {
+            throw new \TypeError(sprintf('Argument 1 passed to %s() must be an instance of %s, %s given.', __METHOD__, TranslatorInterface::class, \is_object($translator) ? \get_class($translator) : \gettype($translator)));
+        }
         $this->translator = $translator;
         $this->translationNodeVisitor = $translationNodeVisitor;
     }
 
+    /**
+     * @deprecated since Symfony 4.2
+     */
     public function getTranslator()
     {
+        @trigger_error(sprintf('The "%s()" method is deprecated since Symfony 4.2.', __METHOD__), E_USER_DEPRECATED);
+
         return $this->translator;
     }
 
@@ -63,7 +71,7 @@ class TranslationExtension extends AbstractExtension
     {
         return array(
             new TwigFilter('trans', array($this, 'trans')),
-            new TwigFilter('transchoice', array($this, 'transchoice')),
+            new TwigFilter('transchoice', array($this, 'transchoice'), array('deprecated' => '4.2', 'alternative' => 'trans" with parameter "%count%')),
         );
     }
 
@@ -101,8 +109,11 @@ class TranslationExtension extends AbstractExtension
         return $this->translationNodeVisitor ?: $this->translationNodeVisitor = new TranslationNodeVisitor();
     }
 
-    public function trans($message, array $arguments = array(), $domain = null, $locale = null)
+    public function trans($message, array $arguments = array(), $domain = null, $locale = null, $count = null)
     {
+        if (null !== $count) {
+            $arguments['%count%'] = $count;
+        }
         if (null === $this->translator) {
             return $this->doTrans($message, $arguments, $domain, $locale);
         }
@@ -110,10 +121,16 @@ class TranslationExtension extends AbstractExtension
         return $this->translator->trans($message, $arguments, $domain, $locale);
     }
 
+    /**
+     * @deprecated since Symfony 4.2, use the trans() method instead with a %count% parameter
+     */
     public function transchoice($message, $count, array $arguments = array(), $domain = null, $locale = null)
     {
-        if (null === $this->translator || !$this->translator instanceof LegacyTranslatorInterface) {
-            return $this->doTransChoice($message, $count, array_merge(array('%count%' => $count), $arguments), $domain, $locale);
+        if (null === $this->translator) {
+            return $this->doTrans($message, array_merge(array('%count%' => $count), $arguments), $domain, $locale);
+        }
+        if ($this->translator instanceof TranslatorInterface) {
+            return $this->translator->trans($message, array_merge(array('%count%' => $count), $arguments), $domain, $locale);
         }
 
         return $this->translator->transChoice($message, $count, array_merge(array('%count%' => $count), $arguments), $domain, $locale);