+++ src/Symfony/Component/Form/Extension/Validator/Type/UploadValidatorExtension.php
@@ -14,6 +14,7 @@ namespace Symfony\Component\Form\Extension\Validator\Type;
 use Symfony\Component\Form\AbstractTypeExtension;
 use Symfony\Component\OptionsResolver\Options;
 use Symfony\Component\OptionsResolver\OptionsResolver;
+use Symfony\Component\Translation\TranslatorInterface as LegacyTranslatorInterface;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
 /**
@@ -25,8 +26,14 @@ class UploadValidatorExtension extends AbstractTypeExtension
     private $translator;
     private $translationDomain;
 
-    public function __construct(TranslatorInterface $translator, string $translationDomain = null)
+    /**
+     * @param TranslatorInterface $translator
+     */
+    public function __construct($translator, string $translationDomain = null)
     {
+        if (!$translator instanceof LegacyTranslatorInterface && !$translator instanceof TranslatorInterface) {
+            throw new \TypeError(sprintf('Argument 1 passed to %s() must be an instance of %s, %s given.', __METHOD__, TranslatorInterface::class, \is_object($translator) ? \get_class($translator) : \gettype($translator)));
+        }
         $this->translator = $translator;
         $this->translationDomain = $translationDomain;
     }