+++ src/Symfony/Component/HttpKernel/EventListener/TranslatorListener.php
@@ -17,6 +17,7 @@ use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\HttpKernel\Event\FinishRequestEvent;
 use Symfony\Component\HttpKernel\Event\GetResponseEvent;
 use Symfony\Component\HttpKernel\KernelEvents;
+use Symfony\Component\Translation\TranslatorInterface as LegacyTranslatorInterface;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
 /**
@@ -29,8 +30,14 @@ class TranslatorListener implements EventSubscriberInterface
     private $translator;
     private $requestStack;
 
-    public function __construct(TranslatorInterface $translator, RequestStack $requestStack)
+    /**
+     * @param TranslatorInterface $translator
+     */
+    public function __construct($translator, RequestStack $requestStack)
     {
+        if (!$translator instanceof LegacyTranslatorInterface && !$translator instanceof TranslatorInterface) {
+            throw new \TypeError(sprintf('Argument 1 passed to %s() must be an instance of %s, %s given.', __METHOD__, TranslatorInterface::class, \is_object($translator) ? \get_class($translator) : \gettype($translator)));
+        }
         $this->translator = $translator;
         $this->requestStack = $requestStack;
     }