+++ src/Symfony/Component/Validator/Context/ExecutionContextFactory.php
@@ -11,6 +11,7 @@
 
 namespace Symfony\Component\Validator\Context;
 
+use Symfony\Component\Translation\TranslatorInterface as LegacyTranslatorInterface;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
@@ -34,8 +35,12 @@ class ExecutionContextFactory implements ExecutionContextFactoryInterface
      *                                               use for translating
      *                                               violation messages
      */
-    public function __construct(TranslatorInterface $translator, string $translationDomain = null)
+    public function __construct($translator, string $translationDomain = null)
     {
+        if (!$translator instanceof LegacyTranslatorInterface && !$translator instanceof TranslatorInterface) {
+            throw new \TypeError(sprintf('Argument 1 passed to %s() must be an instance of %s, %s given.', __METHOD__, TranslatorInterface::class, \is_object($translator) ? \get_class($translator) : \gettype($translator)));
+        }
+
         $this->translator = $translator;
         $this->translationDomain = $translationDomain;
     }