+++ src/Symfony/Component/Validator/Violation/ConstraintViolationBuilder.php
@@ -44,8 +44,14 @@ class ConstraintViolationBuilder implements ConstraintViolationBuilderInterface
      */
     private $cause;
 
-    public function __construct(ConstraintViolationList $violations, Constraint $constraint, $message, array $parameters, $root, $propertyPath, $invalidValue, TranslatorInterface $translator, $translationDomain = null)
+    /**
+     * @param TranslatorInterface $translator
+     */
+    public function __construct(ConstraintViolationList $violations, Constraint $constraint, $message, array $parameters, $root, $propertyPath, $invalidValue, $translator, $translationDomain = null)
     {
+        if (!$translator instanceof LegacyTranslatorInterface && !$translator instanceof TranslatorInterface) {
+            throw new \TypeError(sprintf('Argument 8 passed to %s() must be an instance of %s, %s given.', __METHOD__, TranslatorInterface::class, \is_object($translator) ? \get_class($translator) : \gettype($translator)));
+        }
         $this->violations = $violations;
         $this->message = $message;
         $this->parameters = $parameters;
@@ -142,10 +148,16 @@ class ConstraintViolationBuilder implements ConstraintViolationBuilderInterface
      */
     public function addViolation()
     {
-        if (null === $this->plural || !$this->translator instanceof LegacyTranslatorInterface) {
+        if (null === $this->plural) {
             $translatedMessage = $this->translator->trans(
                 $this->message,
                 $this->parameters,
+                $this->translationDomain
+            );
+        } elseif ($this->translator instanceof TranslatorInterface) {
+            $translatedMessage = $this->translator->trans(
+                $this->message,
+                array('%count%' => $this->plural) + $this->parameters,
                 $this->translationDomain
             );
         } else {